v1.2.0 - Condensed RobotFunctions, RobotFunctionsHandler, and AutoModules into one class: RobotFunctions - Reworked RobotFunctions: Pausing is now a type of stage, and adding a robot function immediately adds it to the queue - Added pausing to Stage: Pausing is now a type of stage – can be set by overriding the isPause method to return true
