+++ TeamCode/src/main/java/util/Stage.java
@@ -2,8 +2,8 @@ package util;
 
 //Used to define exit conditions
 public abstract class Stage {
-    // TODO: Do we really need this class?
     public boolean run(double in) {
         return false;
     } // returns whether it is done or not
+    public boolean isPause() { return false; }
 }
